Conflicts:

	scilab/configure
	scilab/modules/Makefile.in
	scilab/modules/core/includes/stack-c.h
	scilab/modules/core/includes/stack3.h
	scilab/modules/core/includes/version.h
	scilab/modules/core/src/c/stack3.c
	scilab/modules/double/src/c/double.vcproj
	scilab/modules/elementary_functions/sci_gateway/fortran/sci_f_chinesehat.f
	scilab/modules/elementary_functions/sci_gateway/fortran/sci_f_sort.f
	scilab/modules/elementary_functions/src/c/elementary_functions.vcproj
	scilab/modules/gui/src/java/org/scilab/modules/gui/bridge/canvas/ScrolledSwingScilabCanvas.java
	scilab/modules/linear_algebra/src/c/linear_algebra.vcproj
	scilab/modules/optimization/Makefile.in
	scilab/modules/signal_processing/src/c/signal_processing.vcproj
	scilab/modules/sound/Makefile.in

+++ scilab/modules/fileio/src/c/readline.c
@@ -16,6 +16,7 @@
  */
 /*--------------------------------------------------------------------------*/
 #include <stdio.h>
+#include <string.h>
 #include "readline.h"
 #include "filesmanagement.h"
 #include "charEncoding.h"
@@ -105,7 +106,7 @@ int LineRead(FILE *fd,char buf[],int n,int *cnt,int *nr)
   *cnt=count;
   /** Convert from system locale encoding to UTF encoding */
   tmpChar = localeToUTF(buf);
-  *cnt=strlen(tmpChar)+1;
+  *cnt = (int)strlen(tmpChar)+1;
   strcpy(buf,tmpChar);
   return(info);
 }
